Add alt_pdfs functionality; all components done.
A dictionary called `alt_pdfs` is made at the beginning of a run and
passed to each individual in the population. This is now the only
attribute in each column pdf, and is inherited at each mutation.
Possible pdfs to change to are computed when needed by looking
themselves up in the dictionary. This seems easier than trying to figure
out all the possible pdfs each time.