Move seed to fitnesses (#12)
* Remove seed attribute from pdf classes.

* Update fitness process to iterate over seeds

Fitness of an individual is now found by sampling a number of datasets
from the family of datasets with the parameters defined by the
individual's alleles. This all looks correct but the tests are timing
out now.

Will discuss further tomorrow.

* Massively reduce test parameter space.

* Offspring creation and mutation processes updated.

* Add ValueError for small populations

Populations of size 1 or less are not of interest so are excluded for
the sake of making the code look cleaner.

* Add explanation of fitness selection; add ordering

An individual now represents a family of datasets and its fitness is
determined by the fitness of a set of seeded samples from that family.
How to reduce this set of fitness scores down to one is up to the user
and can be done by passing such a function as the `selection` parameter
in a run of the algorithm. By default, the mean of the fitnesses is
taken to be the fitness score of that individual, but realistically, any
function that acts on a list of scores would suffice.

As well as this, the ability to choose how populations are ordered has
been added. Since some fitness functions are to be minimised, a user can
pass a boolean variable to `maximise` indicating whether fitnesses should
be maximised or not.

* Change `selection` to `amalgamation_method`

* Remove `maximise` parameter.

Now the population is sorted by descending fitnesses.

* Slight rewording for `get_fitness` docstring.