The `Family` class, decentralised random states, a better caching framework (#132)
* Implement optimiser class and move to `families` (#122)

* Implement a class for the optimisation process.

* Final clean up. Keeping old tests for scrap.

* Move unit tests outside subdir.

* Opt for `families` over `pdfs` throughout.

* Individual moved to class (#123)

* Implement basic Invidiual class. Fix typos.

* Write `to_history` method over func in `optimiser`

* Write and read methods for individuals. Modularise

* Format codebase.

* Write test for next gen method.

* Complete unit tests for `DataOptimiser`.

* Fix docstrings in optimiser.

* Format codebase.

* Add test to check for reproducibility.

This tests passes but I believe this is only because the Python kernel
is the same in every run. When running tests using the circle example
from the paper (https://github.com/daffidwilde/edo-paper), different
histories are generated when restarting the kernel.

Digging around, it appears this is to do with hashing and the unordered
nature of `set()` (and potentially `dict` although this should not be
the case with Python 3.7+).

In addition to this, I'm not sure how comfortable I am about just
setting the RNG globally with `np.random.seed(global)` and forgetting
it. An attractive solution to that is to create a list of local RNGs
for the individuals in the population that they use for sampling, and
mutating. Perhaps extra RNGs will be needed for fitness calculation and
the crossover process?

If this solution is to be implemented, then classes should be built for
the global algorithm runner and each individual. It may even be the case
that the family classes need looking at.

* Fix kwarg parameter syntax.

* Format codebase with black, isort and flake8.

* Stop using `set` in `do._update_subtypes`

* Add consistency example test.

A test using the circle example to ensure that the EA runs consistently
between runs (i.e. emulating new kernels).

* Remove print statement from test.

* Pickling subtypes, individuals and *families* (#127)

* Implement Family factory handler class.

* Start adding new terminology. [everything broken]

What were families are now distributions. Distribution class is much
cleaner -- is it even necessary anymore?

* Add a repr to the Family class.

* Implement the Family class in individuals.

* Hack failing distribution test.

Since subtypes are not created directly from the distribution classes,
there is no subtype_id in an instance.

* Add subtype family attribute for easy retrieval.

* Implement families in populations.

* Implement families in crossover.

* Update tests for mutation and selection.

* Implement family subtypes in shrink operator.

* Update fitness tests.

* Add Family ability to record all subtypes.

* Begin implementing families in optimiser.

This needs more attention around how individuals are saved.

* Saving histories (individuals and their metadata) (#126)

* Remove `to_dict` method from `Distribution`.

* Fix the circle example test.

* Combine `Individual` to and from file tests.

* Implement new saving strategy.

Families provide their subtypes with a `to_dict` method that records the
name of the subtype's distribution, its subtype ID number and all of its
parameters as a dictionary.

Then when an individual is saved to file, the metadata is serialized as
JSON with these dictionaries. Simultaneously, each family is cached in
the same directory tree.

* Clean up optimiser with new strategy.

The optimiser can now pull usable individuals back in.

* Add a dev environment file.

* Add scipy to environment.yml

* Generalise paths for saving/loading individuals.

* [temporary] Remove parallel writing capabilities.

Because of how individuals and families are being stored, there were
some discrepancies when writing generations to file with multiple
processes. For now, that has been removed. However, the calculation of
fitnesses may still be done in parallel which is where most of the
savings likely come from anyway.

* Format codebase with isort, black and flake8.

* Move to github actions. (#129)

* Fix typo in CI workflow.

* Add test requirements to setup.py and CI workflow.

* Add `dask[dataframe]` to requirements.

* Format codebase with black, flake8 and isort.

* Add fitness to individual (#131)

* Add `random_state` to individual.

* Implement individual random states in operators.

* Implement random states in `DataOptimiser`.

Mutation still needs looking at and so will the distributions.

* Add random_state to distribution instance sampling.

* Add random_state to dist instantiation.

* Implement dist changes to families.

* Add random state (and saving thereof) to family.

* Implement changes in remaining operators.

* Format codebase with black, flake8 and isort.

* Move fitness to the individual rather than cache.

What was I thinking having a global dictionary keeping track of
fitnesses for individuals when I had already implemented the individual
class? Bizarre.

* Remove cache from optimiser.

* Add sampling example used in debugging.

* Localise random states (#130)

* Add `random_state` to individual.

* Implement individual random states in operators.

* Implement random states in `DataOptimiser`.

Mutation still needs looking at and so will the distributions.

* Add random_state to distribution instance sampling.

* Add random_state to dist instantiation.

* Implement dist changes to families.

* Add random state (and saving thereof) to family.

* Implement changes in remaining operators.

* Format codebase with black, flake8 and isort.

* Add tests for coverage on types of random_state.

* Format with black, flake8 and isort.

* Remove deadlines from optimiser tests.

There are some disparities between the performance of the tests in the
CI workflow and local runs.

* More general fitness in final optimiser test.

* Revise the "seed needed" optimiser test.

Add in some assumptions about the minimum number of rows and population
size. Also, change the logic so that the test only fails if *all*
individuals in the populations are the same.

* Update docstrings for core modules.

* Update directory parameter in family test.

* Add in module docstrings for optimiser and version.

* Fix imports for individual.py

* Fix method reading logic in Individual.

* Make custom test parameters for final test.

* Update individual docstring.

* Remove unused `hypothesis.assume` from tests.

* Update docstrings for distributions subpackage.

* Update docstrings of operators subpackage.